[improvements] part 1,  mostly users.component refactored
