+++ src/app/modules/users/models/users.model.ts
@@ -1,19 +1,4 @@
-export interface IUser {
-  _id: string,
-  isActive: boolean,
-  balance: string,
-  picture?: string,
-  age: number,
-  name?: {
-    first: string,
-    last: string
-  },
-  company?: string,
-  email?: string,
-  address: string,
-  tags: string[],
-  favoriteFruit: string,
-}
+import { IUser } from "@shared/models/users.model";
 
 export interface IUsers {
   users: UsersModel;
