+++ src/app/modules/users/services/users.service.ts
@@ -2,7 +2,7 @@ import {inject, Injectable} from '@angular/core';
 import {HttpClient, HttpParams} from "@angular/common/http";
 import {IUsers} from "../models/users.model";
 import {UsersConfig} from "../models/users-config.model";
-import {shareReplay} from "rxjs";
+import {Observable, shareReplay} from "rxjs";
 
 
 @Injectable({
@@ -12,7 +12,7 @@ export class UsersService {
 
   private readonly httpClient = inject(HttpClient)
 
-  fetchUsers(config: UsersConfig) {
+  fetchUsers(config: UsersConfig): Observable<IUsers> {
     let params = new HttpParams();
 
     Object.keys(config.filters).forEach((key) => {
